feat: rename WalletDropdown to Portfolio
