+++ src/components/Banner/UniswapWalletBanner.tsx
@@ -1,8 +1,8 @@
 import { Trans } from '@lingui/macro'
 import { AutoColumn } from 'components/Column'
+import { useWalletDrawer } from 'components/Portfolio'
+import { DownloadButton, LearnMoreButton } from 'components/Portfolio/DownloadButton'
 import Row, { RowBetween } from 'components/Row'
-import { useWalletDrawer } from 'components/WalletDropdown'
-import { DownloadButton, LearnMoreButton } from 'components/WalletDropdown/DownloadButton'
 import { X } from 'react-feather'
 import { useLocation } from 'react-router-dom'
 import { useHideUniswapWalletBanner } from 'state/user/hooks'
