+++ src/components/Popups/TransactionPopup.tsx
@@ -1,8 +1,8 @@
 import { useWeb3React } from '@web3-react/core'
 import Column from 'components/Column'
-import { parseLocalActivity } from 'components/WalletDropdown/MiniPortfolio/Activity/parseLocal'
-import { PortfolioLogo } from 'components/WalletDropdown/MiniPortfolio/PortfolioLogo'
-import PortfolioRow from 'components/WalletDropdown/MiniPortfolio/PortfolioRow'
+import { parseLocalActivity } from 'components/Portfolio/MiniPortfolio/Activity/parseLocal'
+import { PortfolioLogo } from 'components/Portfolio/MiniPortfolio/PortfolioLogo'
+import PortfolioRow from 'components/Portfolio/MiniPortfolio/PortfolioRow'
 import useENSName from 'hooks/useENSName'
 import { useCombinedActiveList } from 'state/lists/hooks'
 import { useTransaction } from 'state/transactions/hooks'
