+++ src/components/Portfolio/AuthenticatedHeader.tsx
