+++ src/components/Portfolio/GitVersionRow.tsx
