+++ src/components/Portfolio/MiniPortfolio/Activity/ActivityRow.tsx
