+++ src/components/Portfolio/MiniPortfolio/ExpandoRow.tsx
