+++ src/components/Portfolio/MiniPortfolio/NFT.tsx
@@ -1,8 +1,8 @@
 import { sendAnalyticsEvent, useTrace } from '@uniswap/analytics'
 import { InterfaceElementName, SharedEventName } from '@uniswap/analytics-events'
 import Column from 'components/Column'
+import { useToggleWalletDrawer } from 'components/Portfolio'
 import Row from 'components/Row'
-import { useToggleWalletDrawer } from 'components/WalletDropdown'
 import { Box } from 'nft/components/Box'
 import { NftCard } from 'nft/components/card'
 import { detailsHref } from 'nft/components/card/utils'
