+++ src/components/Portfolio/MiniPortfolio/PortfolioLogo.test.tsx
