+++ src/components/Portfolio/MiniPortfolio/PortfolioLogo.tsx
