+++ src/components/Portfolio/MiniPortfolio/constants.ts
