+++ src/components/WalletModal/index.tsx
@@ -4,9 +4,9 @@ import { getWalletMeta } from '@uniswap/conedison/provider/meta'
 import { useWeb3React } from '@web3-react/core'
 import { sendEvent } from 'components/analytics'
 import { AutoColumn } from 'components/Column'
+import { useWalletDrawer } from 'components/Portfolio'
+import IconButton from 'components/Portfolio/IconButton'
 import { AutoRow } from 'components/Row'
-import { useWalletDrawer } from 'components/WalletDropdown'
-import IconButton from 'components/WalletDropdown/IconButton'
 import { Connection, ConnectionType, getConnections, networkConnection } from 'connection'
 import { useGetConnection } from 'connection'
 import { ErrorCode } from 'connection/utils'
