+++ src/pages/RemoveLiquidity/index.tsx
@@ -7,7 +7,7 @@ import { BrowserEvent, InterfaceElementName, InterfaceEventName } from '@uniswap
 import { Currency, Percent } from '@uniswap/sdk-core'
 import { useWeb3React } from '@web3-react/core'
 import { sendEvent } from 'components/analytics'
-import { useToggleWalletDrawer } from 'components/WalletDropdown'
+import { useToggleWalletDrawer } from 'components/Portfolio'
 import { useV2LiquidityTokenPermit } from 'hooks/useV2LiquidityTokenPermit'
 import { useCallback, useMemo, useState } from 'react'
 import { ArrowDown, Plus } from 'react-feather'
